# swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "hackerino"
  description: Polls is a simple API allowing consumers to view polls and vote in them.
  license:
    name: MIT
    url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
host: hackerino.herokuapp.com
basePath: /v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /students:
    get:
      description: Returns students list.
      tags:
        - /students
      responses:
        "200":
          description: A students list.
          schema:
            type: array
            items:
              $ref: "#/definitions/Student"
        default:
          description: Unexpected error
    post:
      description: Adds a student.
      tags:
        - /students
      responses:
        "201":
          description: Student created.
        "400":
          description: Bad request.
        default:
          description: Unexpected error
  /students/{studentID}:
    parameters:
      - in: path
        name: studentID
        type: integer
        required: true
    get:
      description: Returns a student by id.
      tags:
        - /students
      responses:
        "200":
          description: A student object.
          schema:
            $ref: "#/definitions/Student"
        "400":
          description: The specified studentID is invalid.
        "404":
          description: A student with the specified studentID was not found.
        default:
          description: Unexpected error
    put:
      description: Updates/creates a student by studentID.
      tags:
        - /students
      responses:
        "200":
          description: Student updated.
        "201":
          description: Student created.
        default:
          description: Unexpected error
    delete:
      description: Removes a student by studentID.
      tags:
        - /students
      responses:
        "204":
          description: Student removed.
        "404":
          description: Could not remove the student with the specified ID (student not found).
        default:
          description: Unexpected error
...
definitions:
...
  Student:
    title: Student
...